+++ crates/orchestrator/src/config.rs
@@ -1,5 +1,6 @@
 use std::sync::Arc;
 
+use aptos_da_client::config::AptosDaConfig;
 use arc_swap::{ArcSwap, Guard};
 use aws_config::SdkConfig;
 use da_client_interface::{DaClient, DaConfig};
@@ -181,6 +182,10 @@ pub async fn build_da_client() -> Box<dyn DaClient + Send + Sync> {
     match get_env_var_or_panic("DA_LAYER").as_str() {
         "ethereum" => {
             let config = EthereumDaConfig::new_from_env();
+            Box::new(config.build_client().await)
+        }
+        "aptos" => {
+            let config = AptosDaConfig::new_from_env();
             Box::new(config.build_client().await)
         }
         _ => panic!("Unsupported DA layer"),